prometheusreceiver: apply suggestions from review
